set fares

+++ src/pages/search-results/SearchResultsOneWay.js
@@ -156,7 +156,7 @@ const SearchResultsOneWay = ({ isMobile }) => {
                           departing_fare = fare.leaving_fare;
                         }
                       })}
-                      $ {parseInt(departing_fare)}
+                      $ {parseFloat(departing_fare)}
                     </Typography>
                   </Stack>
                   <Stack
@@ -182,7 +182,7 @@ const SearchResultsOneWay = ({ isMobile }) => {
                             going_to: flight.going_to,
                             departing_on: departing_on,
                             returning_on: null,
-                            fare: parseInt(departing_fare),
+                            fare: parseFloat(departing_fare),
                             adults: state.formData.adults,
                             kids: state.formData.kids,
                             infants: state.formData.infants,
